Move address helpers to util
