+++ packages/keyring/src/pair/index.ts
@@ -6,9 +6,9 @@ import { Keypair, KeypairType } from '@polkadot/util-crypto/types';
 import { KeyringPair, KeyringPair$Json, KeyringPair$Meta } from '../types';
 import { PairInfo } from './types';
 
+import { encodeAddress } from '@polkadot/util';
 import { naclKeypairFromSeed as naclFromSeed, naclSign, naclVerify, schnorrkelKeypairFromSeed as schnorrkelFromSeed, schnorrkelSign, schnorrkelVerify } from '@polkadot/util-crypto';
 
-import { encodeAddress } from '../address';
 import decode from './decode';
 import encode from './encode';
 import toJson from './toJson';
