+++ packages/util/src/address/decode.spec.ts
@@ -1,9 +1,9 @@
-// Copyright 2017-2019 @polkadot/keyring authors & contributors
+// Copyright 2017-2019 @polkadot/util authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 // @flow
 
-import testingPairs from '../testingPairs';
+import testingPairs from '../../../keyring/src/testingPairs';
 import decode from './decode';
 
 describe('decode', () => {
