+++ packages/util/src/address/encode.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2019 @polkadot/keyring authors & contributors
+// Copyright 2017-2019 @polkadot/util authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
@@ -7,8 +7,11 @@ import { Prefix } from './types';
 // Original implementation: https://github.com/paritytech/polka-ui/blob/4858c094684769080f5811f32b081dd7780b0880/src/polkadot.js#L34
 
 import bs58 from 'bs58';
-import { assert, u8aConcat, u8aToBuffer, u8aToU8a } from '@polkadot/util';
 
+import assert from '../assert';
+import u8aConcat from '../u8a/concat';
+import u8aToBuffer from '../u8a/toBuffer';
+import u8aToU8a from '../u8a/toU8a';
 import defaults from './defaults';
 import sshash from './sshash';
