+++ packages/util/src/address/sshash.ts
@@ -1,9 +1,12 @@
-// Copyright 2017-2019 @polkadot/keyring authors & contributors
+// Copyright 2017-2019 @polkadot/util authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { stringToU8a, u8aConcat } from '@polkadot/util';
+// This is not great - here we now have a dependency from util -> crypto
 import { blake2AsU8a } from '@polkadot/util-crypto';
+
+import stringToU8a from '../string/toU8a';
+import u8aConcat from '../u8a/concat';
 
 const SS58_PREFIX = stringToU8a('SS58PRE');
